multi stuff
built a converter to full paths based on sindbadroot (new). simplified the readConfigFiles (better error handling). adjusted readJsonFile to deal with \g and other special characters. added computer name to stampExperiment. added a flag to the readConfigFiles so that if some fields are missing yield an error (to be discussed if we just set this rule to false).