changing minsb & maxsb functions
minsb and maxsb use the matlab build in min() and max() functions now.
before, minsb(a,b) did not always return the same value as min(a,b). this small difference led to small, but negative values where the minima should be zero, leading to complex numbers for some water states